Modal button fix. Still requires spacing between the login and loout buttons on the homescreen
