+++ client/src/App.css
@@ -5,7 +5,7 @@
 .App-logo {
   /* animation: App-logo-spin infinite 20s linear; */
   height: 200px;
-  margin: auto;
+  margin: 10px auto;
 }
 
 /* .App-header {
