+++ client/src/components/ModalAbout/style.css
@@ -1,0 +1,10 @@
+/* .centerModalAbout {
+
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+    width: 5%;
+
+} */
+
+
