+++ client/src/components/Navbar/style.css
@@ -14,3 +14,7 @@
 #logBtn {
     margin-right: 7em;
 }
+
+#modalBtn {
+    margin-right: 7em !important;
+  }
