+++ package.json
@@ -30,7 +30,7 @@
     "mysql2": "^1.7.0",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
-    "sequelize": "^5.21.5",
+    "sequelize": "^5.21.6",
     "sequelize-cli": "^5.5.1"
   }
 }
